[chore] 서치바 분리
